+++ src/global/validation/pipe/longitutde-validation.pipe.ts
@@ -1,12 +1,12 @@
 import { ArgumentMetadata, Injectable, PipeTransform } from '@nestjs/common';
 import { PlaceResponseCode } from 'src/place/exception/place-response-code';
-import { BaseException } from '../base/base-exception';
+import { BaseException } from '../../base/base-exception';
 
 @Injectable()
-export class LatitudeValidationPipe implements PipeTransform {
+export class LongitudeValidationPipe implements PipeTransform {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   transform(value: number, metadata: ArgumentMetadata) {
-    const isValid = -90 <= value && value <= 90;
+    const isValid = -180 <= value && value <= 180;
 
     if (!isValid) {
       throw BaseException.of(PlaceResponseCode.INVALID_LOCATION);
