+++ src/global/validation/pipe/place-exists-validation.pipe.ts
@@ -1,7 +1,7 @@
 import { ArgumentMetadata, Injectable, PipeTransform } from '@nestjs/common';
 import { PlaceResponseCode } from 'src/place/exception/place-response-code';
 import { PlaceService } from 'src/place/place.service';
-import { BaseException } from '../base/base-exception';
+import { BaseException } from '../../base/base-exception';
 
 @Injectable()
 export class PlaceExistsValidationPipe implements PipeTransform {
