+++ src/global/validation/pipe/sort-condition-validation.pipe.ts
@@ -1,7 +1,7 @@
 import { ArgumentMetadata, Injectable, PipeTransform } from '@nestjs/common';
-import { SortConditon } from 'src/place/dto/request/sort-conditon';
+import { SortConditon } from 'src/place/entity/sort-conditon';
 import { PlaceResponseCode } from 'src/place/exception/place-response-code';
-import { BaseException } from '../base/base-exception';
+import { BaseException } from '../../base/base-exception';
 
 @Injectable()
 export class SortConditionValidationPipe implements PipeTransform {
